First big refactory
This refactor changes the way the date stored, creating Value Objects
for Day, Month and Year.

They work similarly like a linked list, carryng changes to the
next/previus value when needed.

The code still has a lot of duplication, which will be handled in a
future update as we implement more tests.